requirements.txt
streamlit
python-dotenv 